[moonlight] Build mcs.exe as part of the moonlight build
Neither moonlight itself or moonlight applications can be built
with an mcs.exe older than 2.12. As such, to allow users of the moonlight
SDK to build moonlight applications without having a system mono from
git we need to build the compiler ourselves. This will be bundled in the
SDK. We also use it to build all our managed libs as part of the regular
moonlight build.