[mdoc-update] Fix interaction between --delete, -fno-assembly-versions.
`mdoc update --delete` is supposed to remove XML for types and type
members that are no longer present. This works sanely if the
type/member has only one assembly version in it, but once a member is
found in multiple versions, --delete only removes for the assembly
version currently being processed, and thus the member may stay in.

Ordinarily this is sufficient, and is why the HTML output lists the
assembly versions the type/member is found in.

However, then we introduce -fno-assembly-versions, which removes the
version store (if present) and doesn't re-generate. Thus, when both
are present, we would expect types and members to be removed,
irregardless of any pre-existing version information (because that
version information will be removed). This didn't happen. Worse,
because we kept re-processing version information, we'd ~never remove
any members even after repeated invocations, because we weren't
properly skipping the version store. Oops.

Fix the interaction, and Add tests so we ensure this works.