[mdoc-update] Properly implement C#4 dynamic support.
The `dynamic` support in d80fadf was incomplete, in that it would only
find `dynamic` types, and not generic types, e.g.
`Dictionary<string, dynamic>` would be declared as
`Dictionary<string, object>`.

To figure this out, we need to read the
DynamicAttribute.TransformFlags value, which is a bool[] which
specifies which types in the depth-first traversal of the type should
be `dynamic` instead of the actual type, e.g. for
Dictionary<string,object> the bool[] is {false, false, true}.

This requires that we pass a "context" object whenever we have the
"root" of the object, so that we can properly count the depth-first
generic type traversal and generate an appropriate declaration.