Apply patch from Mario Kosmiskaso fix bug 1210: Correctly propagate the Operation Context on async calls
The OperationContext is a [ThreadStatic] variable therefore when a REST call is
made asynchronously the worker thread will not see a context set and will not
expose any data through the context.

A possible solution (which I implemented in the attached patch) is for
ClientRealProxy to set the context in the channel prior to executing the call
(either sync or async). The ClientRuntimeChannel needs to then properly set and
reset the context in the worker thread.

Because WCF operations marked [OperationContract(AsyncPattern = true)] using
IAsyncResult only generate an outgoing call on End<OperationName> the
ClientRuntimeChannel.EndProcess must not reset the context as it's currently
doing. The context can only be reset after calling EndInvoke because the
request will be only generated at this point.