[corlib] On Android, alias stdout/stderr to logcat.
In Mono for Android, stdout and stderr (file descriptors 1 and 2) are
only visible when debugging within MonoDevelop and/or Visual Studio,
which is rather annoying.

This was changed in Mono for Android 1.9.2, by using Console.SetOut()
and Console.SetError() to send stdout/stderr message to both
stdout/stderr and to the Android Debug Log (`adb logcat`), allowing
e.g. Console.WriteLine() output to be viewed outside of a debugger.

Problem: in order to capture user-generated messages at their earliest
point, Console.SetOut()/etc. must be called at the earliest point in
Mono for Android initialization, even if Console.WriteLine() is never
used. Worse, this adds ~180ms of startup overhead on a Nexus One.

The fix here is to move the std*/logcat duplication logic into
mscorlib.dll, so that we can configure things within the
System.Console static constructor. This moves the initialization
penalty onto users of System.Console (instead of all users).