TdsMetaParameter.cs - fixes 699643. Note as Mono.Data.TdsClient and Mono.Data.SybaseClient have both been removed from mono, support for pre 7.0 version of SQL Server isn't required so this should not break applications.
SqlTransaction.cs - Check if the connection is still active before trying to rollback during Dispose. A loss of a connection is possible which result in an exception that will lead to the closing of the session. This will result in the dispose methods being invoked and if there was an active transaction then a rollback will be attempted. However, if the connection is gone we'll end up raising further exceptions about resources already being disposed.