Novell-710512: Fix GetUtcOffset when called with UTC time
Adjust the behavior of the TimeZone.GetUtcOffset() method to
match the documentation. MSDN says: "If the value of the Kind
property is DateTimeKind.Utc, the method returns an offset equal
to TimeSpan.Zero". Previously we returned the difference between
the current time zone and UTC even when the passed in time was
in UTC.

Change-Id: Icaa233f079a74315ccf1ce95787629c37317b412