This is a fix for bug # 853 -- see:  https://bugzilla.xamarin.com/show_bug.cgi?id=853
What was happening was ColumnMappings were never being set, so when another
Update function (which this eventually called) went to map the parameters,
it couldn't find any matching column names, so it set all the parameter values
to null.

This code, or very similar code, appeared in:
 public int Update (DataRow [] dataRows)
 public int Update (DataTable dataTable)

But was left out of:
 public int Update (DataSet dataSet, string srcTable)

This code adds the relevant code into the function it was left out of.

Ideally, longer term, this code would be consolidated into a function which
was commonly called by all 3 update functions, but i didn't want to risk
breaking the other two functions which probably already work fine.