[System] Try to fix Cookies tests
The first part was to try to repro the ordering behavior of MS which seems to be based on the string length of, first, the domain and, then, the path.
Second part was to fix the .local unit test with Dns.GetHostname as some OS (namely Mac) append a '.local' by default to the hostname.