This hides Non-Browsable columns from DataGrid
This addresses at least part of the problem reported here:
https://bugzilla.novell.com/show_bug.cgi?id=322563

When you read a table from XML parent/child table relations are established
by virtue of hidden _Id fields.  Those _Id fields are marked non-browsable
when they are loaded.  This change modifies DataGridTableStyles so that
when it comes across a column that is marked non-browsable, it doesn't make
it browsable (viewable).  That is: It hides the internal use columns.