When reloading a DataGrid rows after, for example, a delete, crash fixed
Anything that would require a change in the number of rows, and trigerred
a DataGrid reload, did not cancel editing of the existing field, so there
was a problem if a row was deleted and you were still editing that row
(had the cursor in it).  I noticed this while testing the following bug report
sample code:

https://bugzilla.novell.com/show_bug.cgi?id=323154

Note: That code crashes if you press spacebar TWICE from my testing in either
.NET or Mono.  However, if you press spacebar ONCE this code behaves also
like .NET in that it will actually delete the row visibly from the screen
and prevent related problems from having an editable field open on a
now-deleted row.

\