[Microsoft.Build.Tasks] Fix several bugs in ResolveAssemblyReference
XBuild was adding both the primary references and nth-order references
to the ResolvedFiles output variable. This meant that dependent assemblies
were being added to the commandline passed to the compiler which caused
all sorts of issues.

Fixes bug https://bugzilla.xamarin.com/show_bug.cgi?id=6708 and adds
tests to cover the changes to ensure they are correct.