[System] Fix Uri.MakeRelativeUri on paths to higher level dir
Given the test case when a relative path should be built from a lower
level dir to a higher level dir, the last bit from the "From-Path" is always
considered to be a file - even if it ends with "/". In contrast, .NET
considers paths ending with "/" as directories and therefore the
resulting path goes one level further up than the resulting path of the
mono implementation. This patch aims to fix that.