[Linq] Return a clearer exception message for Enumerable<T>.First()'s IOE
This method was using the empty constructor (no message) of
InvalidOperationException if no element matched the predicate.
This meant that the message bubbled up would be a bit confusing:
"Operation is not valid due to the current state of the object".

This commit makes Mono match MS.NET by throwing "Sequence contains
no matching element"