[System.Data.SqlClient] Support for "Connection Lifetime" parameter
Adding support for "Connection Lifetime" parameter in connection strings.
This is especially fundamental in this scenario:
- Pooling is ON because the app is in constant reception of requests.
- Loadbalancer is switched to point to a new database.
- Mono needs to discard the connections that are too old (marked by lifetime)
in order to create new ones, that will connect to the new DB.

It includes a unit test in Mono.Data.Tds testing assembly.

Patch contributed under the terms by the X11/MIT license, by:
 James Lewis <james.lewis@7digital.com>
 Andres G. Aragoneses <andres@7digital.com>